feat: move creator interfaces and DraftTagState to commons
Move KMP-compatible creator interfaces to commons module:

- IExpiration: expiration date interface (compose runtime only)
- IZapRaiser: zap raiser amount interface (compose runtime only)
- IMessageField: message field interface (compose foundation text)
- DraftTagState: draft tag state management (compose runtime + coroutines)

All originals replaced with typealiases for backwards compatibility.
