+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/note/creators/expiration/IExpiration.kt
@@ -18,8 +18,11 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.note.creators.draftTags
+package com.vitorpamplona.amethyst.commons.ui.note.creators.expiration
 
-import com.vitorpamplona.amethyst.commons.ui.note.creators.draftTags.DraftTagState as CommonsDraftTagState
+import androidx.compose.runtime.Stable
 
-typealias DraftTagState = CommonsDraftTagState
+@Stable
+interface IExpiration {
+    var expirationDate: Long
+}
